Feature: adding functionality to update BookShelf.
First, the data structure that contains the books was refactored.  Each property of the object is a string that matches the shelf name.  Second, a updateBookShelf function is passed from the App component through to the Book components, since a Book changes the state of the book throughout the app.