Fix: added logic to de-duplicate books in search.
The search API can return the same book multiple times.  In order to duplicate, a Set is created, and if a book's id is already in the set, the book is not added.  Otherwise, a book is added to the result set, and once that loop's completed, the result set's updated.